SubtreeVisibility: Ensure selected elements remain unlocked.
This patch is similar to the focus patch ensuring that we remain
unlocked when any of the elements in our subtree is focused.

It's a bit more complicated since multiple nodes can be selected at once

Note the use of std::set here is because it is both ordered and only
keeps unique elements. The use of it is limited to the local function
which I think is fine.

R=chrishtr@chromium.org

Change-Id: I644c75bc145a0723f868fd7145840e2343837e8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2122857
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755024}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d4035bb954ac86eda978c89c44c88491d49426d5