[Trusted Types] Worker constructor URLs are TrustedScriptURLs
Make URLs used to construct workers TrustedScriptURLs.

Minor rework of how to get the execution context in the generated
bindings, because the old method assumed presence of an "impl" object,
which is only created at the end in case of constructors. So instead,
implement a method with which we can always obtain the execution
context from the v8::FunctionCallbackInfo.

Bug: 1059684
Change-Id: I00e3e52cad5c3736105289a9fda1eeca15fd6f62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2105339
Commit-Queue: Daniel Vogelheim <vogelheim@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Yifan Luo <lyf@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755315}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7bb4923c49b3d122cae306abf81034ea25f4aa4d