SubtreeVisibility: don't unlock in activation (except find-in-page).
This patch removes unlocking for activation. This is instead replaced
with scrolling the element into view for the auto case.

Find-in-page still activates, since it needs further testing and because
it needs to work with hidden-matchable as well as auto.

R=chrishtr@chromium.org, aleventhal@chromium.org

Change-Id: Ice3e488281da6e55bac1451c596e648aaa2b440f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2130680
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755379}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9bee91469c5b3d1adddb7227ad03fb4c7ad0d08d