[EventTiming] Implement EventCounts
This CL adds support for performance.eventCounts, gated behind the
EventTiming flag. EventCounts is maplike and enables computing the
number of input events that have occurred for any of the supported event
types. It is populated on WindowPerformance::RegisterEventTiming so that
it matches when timing with when the PerformanceEventTiming entry is
created for slow events. The spec for this is here:
https://wicg.github.io/event-timing/#sec-event-counts

Bug: 543598

Change-Id: I95bd8b977556557811a1eff8d9fc5d106af0d2b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2125019
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755468}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7d25a90d61ddc8c7475d96f84173f925e1eb2d4e