Let GetBackgroundPaintLocation() reflect the effective paint location
Previously kBackgroundPaintInScrollingContents from
LayoutBoxModelObject::GetBackgroundPaintLocation() meant we should paint
background in the scrolling contents layer only if the scroller was
composited. It would be changed to kBackgroundPaintInGraphicsLayer if
we found the scroller was not composited at multiple places. However,
that didn't work for CompositeAfterPaint.

Now save the the real BackgroundPaintLocation in LayoutObject
bitfields (in place of previous PreviousBackgroundPaintLocation), and
invalidate background when it changes.

This fixes CompositeAfterPaint issue that background was painted in
scrolling contents when the scroller was not composited.

This also improves raster invalidation in some cases.

Change-Id: Ide87d31f63eb14de909edd65aff1e88f1b44d984
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127991
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755478}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7fc1f50f7522f2390eb845f60cd314e6217b997