[AdTagging] Enable top of stack scanning for async ad tagging
What: Scan the top of the stack as well for async ad tagging.

Why: Scanning just the top of the stack misses out on a number of
cases.

How: Scanning the top of the stack was previously slow, but this CL
switches the call from SourceLocation::Capture to
v8::StackTrace::CurrentStackTrace which is 10x faster.

Bug: 851531
Change-Id: I2cdbb51a256b6253ba99669a8a2ff0740480b3f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2131032
Commit-Queue: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755588}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60f7e667e5ea3d885fd943d14d638789953e4ce6