DedicatedWorker should reject script w/o COEP when its parent has COEP
[1] is changing how COEP is set to (dedicated) worker top-level scripts.
Before the change it inherited the parent policy. After the change it
gets the policy from the HTTP headers in the response, but the loading
is blocked when the parent has COEP: require-corp but the child doesn't
have COEP: require-corp.

As described in [2], this can be solved by a project called
PlzDedicatedWorker, but it will take time to ship it. As a short-term
fix, this implements the blocking part. No reporting is implemented.

This fix is important because this is a breaking change. We would like
to land this by M83.

1: https://github.com/WICG/cross-origin-embedder-policy/pull/13
2: https://docs.google.com/document/d/1eB34zYLuQHshN_Tt8BbLSehHyvxqWupVdB09CCbn7g4/edit

Bug: 1064920
Change-Id: I55fa951f18692d642c747cff168d8239de69efb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2126652
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755782}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8007a3d5e289c8b8d31dad0a0ccae09ead4556d2