[RTCInsertableStreams] Make RTCEncodedVideoFrame serializable
The use case to be supported is efficiently running Streams of
RTCEncodedVideoFrame in a Worker. In this case, RTCEncodedVideoFrame
are serialized to the Worker.

Serialization is supported only in the same process as a shallow copy
where the underlying webrtc frame is shared among the original and
the deserialized object.
See https://github.com/alvestrand/webrtc-media-streams/issues/5

Bug: 1052765
Change-Id: I1e2d9bcd2e24ef6aa6a7688f34283a81e6d23172
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083297
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 283079bc066cc19c0d0d186500697ca95ca03c3f