[RTCInsertableStreams] Make RTCEncodedAudioFrame serializable
The use case to be supported is efficiently running Streams of
RTCEncodedVideoFrame in a Worker. In this case, RTCEncodedVideoFrame
are serialized to the Worker.

Serialization is supported only in the same process as a shallow copy
where the underlying webrtc frame is shared among the original and
the deserialized object.
See https://github.com/alvestrand/webrtc-media-streams/issues/5

Bug: 1052765
Change-Id: Ie236cb3d0e0537ac55b07f030eee028cb5706217
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2132156
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755880}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7415e4fb52b66c115395e0b08eb79c7d10326130