Update build.gn to be in sync after patches.
We've been using a manually modified build.gn file because of the
androidx migration. If you just run fetch_all.py, it should generally
do nothing. When updating a library, users have to manually patch
in the differences to not override the manual changes. These are
changes that are not part of the manual modifications that haven't
been incorporated into the build.gn

It looks like the manual modifications may be in for longer than we'd
like as weblayer/public can't update and a few third_party code won't
migrate soon, so will look into making the local modifications
not need a manual patch.

Additionally, build.gradle was downloading espresso 3.2, but the DEPS
is pointing to 3.1, so sync up those numbers.

Change-Id: I3cbad618e9db2a46d06624287510963756c0d622
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2133266
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: benjamin joyce <bjoyce@chromium.org>
Auto-Submit: benjamin joyce <bjoyce@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#755896}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b0ab5b8b4d38e76d0880d14f3bcd449012d7991e