Fix computed style for outline-offset
Computed style for outline-offset was not returning the proper value
if "outline-style" was "none".
This was due to a check in ComputedSyle::OutlineOffset()
that this patch removes.

This code could be there because this is needed for "outline-width"
as specified in the spec [1], however there's nothing about this
on the "outline-offset" property.

This patch modifies outline-offset-computed.html, as it had
some extra lines (similar to what's needed for "outline-width")
to make it work in Chromium (and WebKit https://webkit.org/b/209913)
due to this bug.

[1] https://drafts.csswg.org/css-ui/#outline-props

Bug: 1005372
Test: css/css-ui/parsing/outline-offset-computed.html
Change-Id: Ief8d4fe22c462ec2bb493b20b1b22b8d0797e1ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134133
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Manuel Rego <rego@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#755943}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7a7a80fee0878fee8e6bba1e384c1676de49b61d