Refactor handling of surrogates in StyleCascade (again)
This CL does some refactoring necessary to reduce the noise in
subsequent CLs.

Instead of returning a value from ResolveIfSurrogate indicating
whether the surrogate should be skipped or not, we first determine if
the property is a surrogate, and if so, call ApplySurrogate, and deal
with any skip/apply logic inside that function.

This CL should have no externally observable behavior change, though it
does change one detail internally in the StyleCascade: the now-removed
ResolveSurrogate function called LookupAndApply on the original property
as a way to ensure that this property was marked as applied. The new
function (ApplySurrogate) simply marks it as such, effectively skipping
it. For symmetry this also adds MarkUnapplied, although this function
is unused at the moment. (It will be needed as part of the work on
'revert').

Bug: 579788
Change-Id: Ic10b341f9d6df3cfaa983e10b67cad0cace48a5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2128073
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#756019}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec7d74ea676ddd86276c0944f5369858b986a5dd