Add perf test for converting to blob in worker
We currently test toblob, but only on the main thread. It looks as
though we can be at the mercy of the scheduler here, so it would be
worthwhile to at least keep track of it.

See: crbug.com/1053477

Bug: 1065481
Change-Id: I42ee1f9ae20dbedea45952e7a79ada10d7f9b19b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124971
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Yi Xu <yiyix@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#756038}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d6574c47c31631f9edf2160dde0ee87f5351ae8