Remove DisableCompositingQueryAsserts from PaintLayer codepaths
These three locations in code all have a DisableCompositingQueryAsserts
disabler in order to query some compositing state. However, none of
these operations are necessary in CompositeAfterPaint.

In PaintLayer destructor, we won't have a composited layer mapping to
clear in CAP. Enforce callers of ClearCompositedLayerMapping to have
ensured that HasCompositedLayerMapping is true, clean up an unnecessary
CAP check and more tightly scope the DisableCompositingQueryAssert to
this function.

PaintLayer::StyleDidChange queries compositing state to determine
whether or not to call SetNeedsCompositingInputsUpdate. That method is
a no-op in CAP.

PaintLayer::DirtyStackingContextZOrderLists queries compositing state
in order to mark the AncestorStackingContext's CompositedLayerMapping
as needing a graphics layer update, which also is not needed in CAP.

Bug: 1007989
Change-Id: I552b348a3bc4d6da196eaad90b05e25a6b470c97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129090
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#756315}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa6b82a589be3353eeebd5aa4926f9c82efa079f