[LayoutNG] Large legend small border fragmentation
When the legend is larger than the fieldset border-top, the fieldset's
border box gets pushed down. In the case of fragmentation, allow
this adjustment if the legend fits inside the first fragment.
Otherwise, don't adjust the fieldset border block-start offset.

This is similar to how fragmentation was handled in the case where the
legend is smaller than the border-top:
https://chromium-review.googlesource.com/c/chromium/src/+/2125230

Tests were not added at this time for the following reasons:
1. It currently isn't possible to test where the fieldset border
starts using unit tests.
2. Paint for legend fragmentation isn't currently working, so testing
this via web tests was also not possible.

Bug: 875235
Change-Id: I9594ab35bce23a4e9dd1a3aa2e7d913900ec14b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2131348
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#756351}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6f64a55f0e605684c9d53b06e6b4fd7cde2c200d