[css-flexbox] Restore flexitem-stretch-image.html out of WPT
The original WPT migration of flexitem-stretch-image.html has
caused a test failure on Firefox. The test will be updated to
work on Firefox. So, the migrated WPT test needs to be kept
in order to make Firefox have the test. Instead, the original
test is restored in the css3/flexbox to avoid regressions from
the existing behavior. Additionally, this CL adds the test
as flex-item-stretch-image.html to avoid the same naming
confusion with the WPT test.

Bug: 1063749
Change-Id: I2633293f05c3a584e73b6594d12a7c85abe39103
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135659
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#756943}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 08e3d24ffe0043c51b3f3b7c789b89f2be647908