Use outline-offset for default focus ring.
Prior to this CL, the outline-offset was *ignored when drawing the
default focus ring (see AdjustedFocusRingOffset):
- for FormControlsRefresh, GetDefaultOffsetForFocusRing was used instead
- on Windows the value was completely ignored
- on Mac the value was used

This CL updated the behavior so that the default focus ring respects the
value of outline-offset.
Because of this, GetDefaultOffsetForFocusRing is not needed anymore
since now we can directly update the outline-offset to correctly
reflect what we are painting.
Another difference in behavior is that the offset is not going to be
zoom independent anymore.

This fix is just for FormControlsRefresh (the feature is enabled
on all platforms except Android) to minimize regression risks.

The fix is validated by: focus-ring-outline-offset.html

Bug: 945181
Change-Id: I2cff82556571d9f541b95f8cece9b7c3be1191c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138156
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#756951}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f152cbdd19975d8ac11da4c7b66b1f2924f855e7