Add core probe to set DevToolsId and DevToolsToken
The probe sets DevToolsId and DevToolsToken, so that they can be sent
to the DevTools frontend whenever an issue is detected in the mixed
content check.

Previously DevToolsId and DevToolsToken were set in PrepareRequest(),
which is executed only after the check for mixed content has already
happened. When generating a DevTools issue in the mixed content check,
we need DevToolsId and DevToolsToken to be available. Therefore this
CL adds a core probe, which allows setting DevToolsId and DevToolsToken
earlier.

Bug: chromium:1053468

Change-Id: I42f50ae34dad7a09095447fb398f2af2cc23bd0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2117735
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757049}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6bb4c8c223311db9dcdcdbe7a7c1b8ae9ccb04e3