Retain Owner from HeapMojoReceiverSet::Wrapper
There is a chance that the HeapMojoReceiverSet::Wrapper may outlive the Owner in case of a ConservativeGC.
The stack scanning may create a root to HeapMojoReceiverSet::Wrapper, prolonging its life.
To make sure they die together in the same GC cycle, we will make the Wrapper retain the Owner.
So in case HeapMojoReceiverSet::Wrapper is kept alive, the Owner is guaranteed to be alive too.

For regular mojo::ReceiverSet, the reference to the Interface implementation is saved in the mojo::ReceiverSet as a raw pointer.
Hence the convention is for the Interface implementation to be the owner.
And since the owner for a HeapMojoReceiverSet cannot change during its lifetime, this CL changes where the Interface is supplied, from Add() to the constructor.

Bug: 1067476
Change-Id: I12f83e922b624a58eb2cfc8e2bfd05228d055f5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2139953
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757344}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 485438fa50a44600fc602865b9ddd10de23f2b7a