appcache: remove token expires from entry in chrome://appcache-internals
This is a followup to:
https://chromium-review.googlesource.com/c/chromium/src/+/2132592

That CL showed token_expires on both entries and groups.  It turns out
that it should be sufficient to check only the group, so the UI on entry
has been removed to avoid confusing developers.

Bug: 582750
Change-Id: Ib575241a8f1988ce40b87f9568ff2e4bbc7310a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2140104
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757464}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 891d35372b5fc7a94f60f2b90a455a838cbd2682