[css-filters] Migrate effect-reference-on-span.html test to WPT
Migrate this test out of third_party/blink/web_tests/css3/filters
and into the WPT-specific directory, adding links to the relevant
specs and a test assertion describing its purpose.

Additionally, this test has been slightly modified to become a
reftest instead of a crash test, so that we not only check that
applying a CSS filter to a <span> element does not crash, but
also that it yields the expected results.

Bug: 1063749
Change-Id: Icd7234480f6f64a15c4af0ae17c8a19199a12080
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2141930
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#757473}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ca107e50b7c67b75d41525c3ddaa3a1e1eb71fea